requirements.txt
pandas
numpy
gradio
google-generativeai
faiss-cpu
sentence-transformers
